+++ R/portion.R
@@ -203,6 +203,10 @@ portion_df <- function(data,
                        minimum = 30,
                        as_percent = FALSE) {
 
+  if (!"data.frame" %in% class(data)) {
+    stop("`portion_df` must be called on a data.frame")
+  }
+
   if (data %>% select_if(is.rsi) %>% ncol() == 0) {
     stop("No columns with class 'rsi' found. See ?as.rsi.")
   }
@@ -240,7 +244,7 @@ portion_df <- function(data,
 
   res <- bind_rows(resS, resI, resR) %>%
     mutate(Interpretation = factor(Interpretation, levels = c("R", "I", "S"), ordered = TRUE)) %>%
-    tidyr::gather(Antibiotic, Percentage, -Interpretation, -data.groups)
+    tidyr::gather(Antibiotic, Value, -Interpretation, -data.groups)
 
   if (!translate_ab == FALSE) {
     if (!tolower(translate_ab) %in% tolower(colnames(AMR::antibiotics))) {